+++ builder_float.go
@@ -89,12 +89,6 @@ func (b *FloatBuilder[T]) WithMaximum(v T) *FloatBuilder[T] {
 	return b
 }
 
-// SeeAlso creates a relationship between this variable and those used by i.
-func (b *FloatBuilder[T]) SeeAlso(i Input, options ...SeeAlsoOption) *FloatBuilder[T] {
-	seeAlsoInput(&b.spec, i, options...)
-	return b
-}
-
 // Required completes the build process and registers a required variable with
 // Ferrite's validation system.
 func (b *FloatBuilder[T]) Required(options ...RequiredOption) Required[T] {