+++ relationship_seealso.go
@@ -2,20 +2,26 @@ package ferrite
 
 import "github.com/dogmatiq/ferrite/variable"
 
+// SeeAlso is an option that adds i to the "see also" section of the generated
+// documentation.
+func SeeAlso(i Input, options ...SeeAlsoOption) interface {
+	DeprecatedOption
+	RequiredOption
+	OptionalOption
+} {
+	return option{
+		Input: func(cfg *inputConfig) {
+			for _, v := range i.variables() {
+				seeAlso(cfg.Spec.Peek(), v.Spec(), options...)
+			}
+		},
+	}
+}
+
 // SeeAlsoOption changes the behavior of a call the SeeAlso() method of the
 // various builder implementations.
 type SeeAlsoOption interface {
 	applySeeAlsoOption(*variable.RefersTo)
-}
-
-func seeAlsoInput(
-	from variable.SpecBuilder,
-	to Input,
-	options ...SeeAlsoOption,
-) {
-	for _, v := range to.variables() {
-		seeAlso(from.Peek(), v.Spec(), options...)
-	}
 }
 
 func seeAlso(